+++ tests/transitions/HotToColdTransitionTest.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using werkbank.models;
+using werkbank.operations;
 using werkbank.repositories;
 using werkbank.services;
 using werkbank.transitions;
@@ -38,6 +39,31 @@ namespace tests.transitions
             Assert.IsTrue(File.Exists(coldMetaFile));
             Assert.IsTrue(File.Exists(Path.Combine(werk.GetDirectoryFor(WerkState.Cold), "my-content.txt")));
             Assert.AreEqual(JsonConvert.SerializeObject(werk), File.ReadAllText(coldMetaFile));
+        }
+
+        [TestMethod]
+        public void HiddenStillHidden()
+        {
+            Werk werk = Util.CreateDummyWerk(EnvironmentRepository.Environments[3], WerkState.Hot);
+
+            Directory.CreateDirectory(Path.Combine(werk.CurrentDirectory, "s1"));
+            DirectoryInfo subDir2 = Directory.CreateDirectory(Path.Combine(werk.CurrentDirectory, "s2"));
+            string file1 = Path.Combine(werk.CurrentDirectory, "f1.txt");
+            File.WriteAllText(file1, "hi");
+
+            Hide.Perform(subDir2.FullName);
+            Hide.Perform(file1);
+
+            HotToColdTransition transition = new();
+            Batch batch = transition.Build(werk);
+            Util.WorkOffBatch(batch);
+
+            Assert.IsNull(werk.TransitionType);
+            Assert.IsFalse(Directory.Exists(werk.GetDirectoryFor(WerkState.Hot)));
+            Assert.IsTrue(Directory.Exists(werk.GetDirectoryFor(WerkState.Cold)));
+            Assert.IsFalse(new DirectoryInfo(Path.Combine(werk.GetDirectoryFor(WerkState.Cold), "s1")).Attributes.HasFlag(FileAttributes.Hidden));
+            Assert.IsTrue(new DirectoryInfo(Path.Combine(werk.GetDirectoryFor(WerkState.Cold), "s2")).Attributes.HasFlag(FileAttributes.Hidden));
+            Assert.IsTrue(File.GetAttributes(Path.Combine(werk.GetDirectoryFor(WerkState.Cold), "f1.txt")).HasFlag(FileAttributes.Hidden));
         }
 
         [TestMethod]